rm typesaurus and use converter
