+++ functions/skeet/src/models/lib/vertex/streamResponse.ts
@@ -1,22 +1,6 @@
-import { db } from '@/index'
 import { ChunkedStream } from '@/lib/chunk'
-import { createFirestoreDataConverter } from '@/models/converters'
-import {
-  VertexChatRoomMessage,
-  VertexChatRoom,
-  VertexChatRoomCN,
-  UserCN,
-  VertexPromptCN,
-  VertexChatRoomMessageRole,
-} from '@/models'
 
-export const streamResponse = async (
-  response: any,
-  res: any,
-  user: any,
-  req: any,
-  vertexChatRoomData: VertexChatRoom,
-) => {
+export const streamResponse = async (response: any, res: any) => {
   const stream = new ChunkedStream(response)
 
   stream.on('data', async (chunk) => {
@@ -24,31 +8,11 @@ export const streamResponse = async (
       res.setHeader('Content-Type', 'application/json')
     }
     console.log(`chunk: ${chunk}`)
-    res.write(JSON.stringify({ text: chunk.toString() }))
+    await res.write(JSON.stringify({ text: chunk.toString() }))
   })
 
   stream.on('end', async () => {
-    const messageCollectionRef = db
-      .collection(
-        `${UserCN}/${user.uid}/${VertexChatRoomCN}/${vertexChatRoomData.userId}/${VertexPromptCN}/${req.body.vertexPromptId}/VertexChatRoomMessage`,
-      )
-      .withConverter(createFirestoreDataConverter<VertexChatRoomMessage>())
-
-    const messageBody = {
-      vertexChatRoomId: vertexChatRoomData.userId,
-      role: VertexChatRoomMessageRole.USER,
-      content: req.body.content,
-    }
-    await messageCollectionRef.add(messageBody)
-
-    const messageResBody = {
-      vertexChatRoomId: vertexChatRoomData.userId,
-      role: VertexChatRoomMessageRole.AI,
-      content: response,
-    }
-    await messageCollectionRef.add(messageResBody)
-
-    res.end()
+    res.end('Stream done')
   })
 
   stream.on('error', async (e: Error) => {
