+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "skeet-app",
-  "version": "1.2.0",
+  "version": "1.3.0",
   "description": "Skeet Framework Boilerplate",
   "author": "ELSOUL LABO B.V.",
   "license": "Apache-2.0",
